% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_tree_with_presents_function.R
\name{draw_tree_with_presents}
\alias{draw_tree_with_presents}
\title{A X-mas tree function}
\usage{
draw_tree_with_presents(number_of_presents)
}
\arguments{
\item{number_of_presents}{how many you want under the tree}
}
\description{
This function allows you to draw a tree with some baubles and presents
}
\examples{
draw_tree_with_presents(34) ## --> look at your tree in the plot-pane :D
}
\keyword{}
